pages_spellchecker/requirements.txt
pyspellchecker==0.7.2
